backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
pymupdf==1.27.2.2
unstructured[docx]==0.16.25
python-docx==1.1.2
chromadb==1.4.1
sentence-transformers==3.0.0
ollama==0.2.0
spacy==3.8.2
rank-bm25==0.2.2
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
aiofiles==23.2.1
python-jose==3.3.0
httpx==0.27.0
